+++ monitor-services/bgpstreamlivetap/core/bgpstreamlive.py
@@ -501,9 +501,7 @@ class BGPStreamLiveDataWorker:
                                             serializer="ujson",
                                         )
                                 else:
-                                    log.warning(
-                                        "Invalid format message: {}".format(msg)
-                                    )
+                                    log.debug("Invalid format message: {}".format(msg))
                             except BaseException:
                                 log.exception(
                                     "Error when normalizing BGP message: {}".format(msg)